Split functions into separate scripts
